fix(plugin-openclaw): review round 30 for #3074 — bound the batch flush; extract hook-field readers
- the batch flush POST gets at most half the remaining lifecycle deadline,
  like the support probe: its failure path is flushIndividually, and a batch
  that stalled having spent everything left those singular flushes the 1ms
  floor and buffered the ended session's turns anyway
- moves the pure hook-payload readers (sessionKeyFrom, lifecycleSessionKeyFrom,
  recallQueryFrom, cwdFrom, readContextComposition) into
  delegate-hook-fields.ts: delegate-runtime.ts was sitting on the 1200 LOC cap,
  and the ratchet asks for extraction rather than shrinking comments

Part of #3073
