+++ .changeset/openclaw-discovery-runtime-delegate-tools.md
@@ -18,6 +18,6 @@ OpenClaw 2.0 delegate-mode fixes for hosts that register in the discovery modes.
 - Delegate mode registers daemon-backed `memory_search` (through the capability's search manager, so scoping and authorization match the host's own memory search) and `memory_get` (`GET /engram/v1/memories/{id}`) tools when the host exposes `registerTool`.
 - Delegate mode does NOT register the OpenClaw 2.0 `registerMemoryPromptPreparation` step: it runs before `before_prompt_build` and is handed no session namespace, so it could only recall a previous scope. `before_prompt_build` injects this turn's recall in the session's bound scope.
 - The recall query is the current turn (`event.prompt`, envelope-cleaned) capped at 1500 characters, keeping the tail, so a host that hands the hook an assembled prompt cannot 413 the daemon's body limit.
-- The lifecycle flush's batch-support probe (`/engram/v1/capabilities`) is capped at half the remaining deadline, so a cold probe that stalls cannot leave the per-namespace fallback flush without usable time.
+- The lifecycle flush's batch-support probe (`/engram/v1/capabilities`) AND the batch flush request are each capped at half the remaining deadline, so neither can stall away the budget the per-namespace fallback flush needs.
 - `agent_end` no longer blocks on the observe POST: the turn capture is detached and chained per session, bounded by half the flush timeout so that a flush for the same session (which waits behind it for at most that long) can never overtake one observe and always keeps the other half for its own requests. When a whole QUEUE of observes outlasts that wait, the flush proceeds and a follow-up flush is chained behind the queue, so late turns are never left buffered past `session_end`. The follow-up drains a bounded number of rounds and releases its guard before flushing, so a session that keeps observing past the cutoff gets another follow-up behind the remainder rather than leaving it buffered. It drains the whole queue (including turns observed after the hook returned, which the daemon buffers per session) and resolves its scope from the session's bindings rather than the ended event's captured metadata.
 - A `REMNIC_HOST` that names one of this machine's own interface addresses (a NIC or VIP) is dialed through loopback, like a wildcard bind; a gateway fetch to such an address has been observed to hang.
